+++ scripts/module.js
@@ -2,6 +2,7 @@ import { MODULE_ID, SOURCES } from "./const.js";
 import { messageItemHasRollOption } from "./helpers.js";
 import { extractDCValueRegex, isIncarnate } from "./incarnate.js";
 import { isMechanic, setMechanicRelevantInfo } from "./specificClasses/mechanic.js";
+import { isSummoner, setSummonerRelevantInfo } from "./specificClasses/summoner.js";
 import { setupSettings } from "./settings.js";
 import { getSpecificSummonDetails } from "./specificSummons.js";
 import { handleUpdateMessage } from "./updateMessage.js";
@@ -21,12 +22,18 @@ Hooks.once("ready", async function () {
   Hooks.on("createChatMessage", async (chatMessage, _info, userID) => {
     if (userID !== game.user.id) return;
 
+    const isBindHeroicSpiritSuccess = isBindHeroicSpiritHit(chatMessage);
+
+    if (chatMessage.isRoll && !isBindHeroicSpiritSuccess) return;
+
     // Handle Specific Case Bind Heroic Spirit
-    const itemUuid = isBindHeroicSpiritHit(chatMessage)
+    const itemUuid = isBindHeroicSpiritSuccess
       ? SOURCES.NECROMANCER.BIND_HEROIC_SPIRIT_STRIKE
       : chatMessage?.item?.sourceId;
 
     if (!itemUuid) return;
+
+    if (chatMessage?.flags?.pf2e?.appliedDamage) return;
 
     // TODO handle Incarnate spells at a later date
     //if (!chatMessage?.flags?.pf2e?.origin?.rollOptions?.includes("summon")) return;
@@ -42,15 +49,29 @@ Hooks.once("ready", async function () {
     if (isMechanic(chatMessage)) {
       setMechanicRelevantInfo(summonerActor, spellRelevantInfo);
     }
+    if (isSummoner(chatMessage)) {
+      setSummonerRelevantInfo(summonerActor, spellRelevantInfo);
+    }
+    if (itemUuid === SOURCES.COMMANDER.PLANTED_BANNER) {
+      spellRelevantInfo.int = summonerActor.system.abilities.int.mod;
+    }
 
-    let summonDetailsGroup = getSpecificSummonDetails(itemUuid, spellRelevantInfo)
+    let summonDetailsGroup = await getSpecificSummonDetails(itemUuid, spellRelevantInfo)
     if (!summonDetailsGroup) {
       summonDetailsGroup = getTraditionalSummonerSpellDetails(itemUuid, spellRank);
     }
 
-    const summonType = isMechanic(chatMessage) ? "mechanic" : (messageItemHasRollOption(chatMessage, "thrall") ? "thrall" : "summon")
+    const summonType = getSummonType(chatMessage);
     await summon(summonerActor, itemUuid, summonType, summonDetailsGroup);
   });
 });
 
-
+function getSummonType(chatMessage) {
+  if (isMechanic(chatMessage))
+    return "mechanic";
+  if (messageItemHasRollOption(chatMessage, "thrall"))
+    return "thrall";
+  if (isSummoner(chatMessage))
+    return "summoner";
+  return "summon";
+}