+++ scripts/updateMessage.js
@@ -5,6 +5,46 @@ const sampleData = {
 };
 
 const updateData = {
+    '1.10.0': {
+        new: [
+            { item: "Support for the following:", feat: true },
+            {
+                children: [
+                    "`SF2e - Summon Robot` (🖥️ @Sasane)"
+                ]
+            },
+        ],
+        update: [
+            { item: "Fix for `Commander - Plant Banner`" },
+            { item: "Fixed `Necromancer - Summon Thrall` activating on `attack rolls` and `Damage Applied` messages" },
+            { item: "Added CLI to clear up the packs thanks to (@Vauxs)" },
+            { item: "Updated french translation (@rectulo, @Sasane)" },
+        ],
+    },
+    '1.9.0': {
+        new: [
+            { item: "Support For Commander - Plant Banner", feat: true },
+            {
+                children: [
+                    "This is only the start of automation (IE handles the base case but not other effects modifying the banner itself)",
+                    "Those may / hopefully will be handled at a later time",
+                    "Also bless @Vauxs for thinking of crosshairParams to begin with"
+                ]
+            },
+        ],
+        update: [
+            { item: "Fix for missing logic to handle `Bind Heroic Spirit` (@Loki123)" },
+            { item: "Fix for `Light` spell not properly handling higher level casts" }
+        ],
+    },
+    '1.8.0': {
+        new: [
+            { item: "Added Support for the following" },
+            { children: ["Conglomerate of Limbs", "Manifest Eidolon (🖥️ @Sasane)"] },
+        ],
+        update: [
+        ],
+    },
     '1.7.0': {
         new: [
             { item: "Living Graveyard - Movement Summon", feat: true },